+++ CHANGES.md
@@ -7,7 +7,8 @@ Beta Releases
 ### b25 - 2014-02-03
 * Breaking changes:
   * The `Viewer` constructor argument `options.fullscreenElement` now matches the `FullscreenButton` default of `document.body`, it was previously the `Viewer` container itself. 
-  
+* Added `ScreenSpaceCameraController.enableAnyInputs` to fix issue with inputs not being restored after overlapping camera flights.
+
 ### b24 - 2014-01-06
 
 * Breaking changes: